requirements.txt
anytree
asyncio
beautifulsoup4
bencoding
cfscrape
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
psutil
pybase64
python-dotenv
python-telegram-bot
requests
telegraph
tenacity
xattr
